+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [unreleased]
+
+- added '--longhorn-custom-settings' option to env/setup.sh to install longhorn
+  using custom settings.
+
 ## [0.4.0] - 2022-09-01
 
 ### Added